Adjust build commands, fix tsc errors
