+++ .github/workflows/deploy-staging.yaml
@@ -13,7 +13,7 @@ jobs:
       - name: Package
         run: |
           npm ci
-          npm run build
+          npm run build:export
       - uses: actions/upload-artifact@v2
         with:
           name: dist
