+++ .github/workflows/test.yml
@@ -96,7 +96,7 @@ jobs:
           CYPRESS_USER_NAME: ${{ secrets.CYPRESS_USER_NAME }}
           CYPRESS_USER_PASSWORD: ${{ secrets.CYPRESS_USER_PASSWORD }}
         with:
-          build: npm run build
+          build: npm run build:export
           start: npx http-server -a localhost -p 3000 dist --proxy http://localhost:3000/\?
           wait-on: http://localhost:3000/
           wait-on-timeout: 120
