+++ src/@next/hooks/useCheckoutStepState.ts
@@ -2,8 +2,10 @@ import { useEffect, useState } from "react";
 
 import { IItems, ITotalPrice } from "@saleor/sdk/lib/api/Cart/types";
 import { ICheckout, IPayment } from "@saleor/sdk/lib/api/Checkout/types";
-import { CheckoutStep } from "@temp/core/config";
-import { checkIfShippingRequiredForProducts } from "@utils/core";
+import {
+  CheckoutStep,
+  checkIfShippingRequiredForProducts,
+} from "@pages/CheckoutPage/utils";
 import { isPriceEqual } from "@utils/money";
 
 interface StepState {
