+++ src/@next/pages/CheckoutPage/subpages/CheckoutAddressSubpage.tsx
@@ -15,9 +15,12 @@ import { ShopContext } from "@temp/components/ShopProvider/context";
 import { commonMessages } from "@temp/intl";
 import { IAddress, IFormError } from "@types";
 import { filterNotEmptyArrayItems } from "@utils/misc";
-import { CheckoutStep } from "@temp/core/config";
 
-import { SubpageBaseProps, SubpageCompleteHandler } from "../utils";
+import {
+  CheckoutStep,
+  SubpageBaseProps,
+  SubpageCompleteHandler,
+} from "../utils";
 
 const CheckoutAddressSubpageWithRef: RefForwardingComponent<
   SubpageCompleteHandler,
