+++ src/@next/pages/CheckoutPage/utils.tsx
@@ -3,16 +3,75 @@ import { FormattedMessage } from "react-intl";
 import { NextRouter } from "next/router";
 import { ParsedUrlQueryInput } from "querystring";
 
-import type { CartSummaryProps } from "@components/organisms";
+import type { CartSummaryProps } from "@components/organisms/CartSummary";
 import type { IItems } from "@saleor/sdk/lib/api/Cart/types";
 import { checkoutMessages } from "@temp/intl";
-import {
-  CheckoutStep,
-  CheckoutStepDefinition,
-  CHECKOUT_STEPS,
-} from "@temp/core/config";
-import { checkIfShippingRequiredForProducts } from "@utils/core";
+
 import { paths } from "@paths";
+
+export const checkIfShippingRequiredForProducts = (items?: IItems) =>
+  items?.some(({ variant }) => variant.product?.productType.isShippingRequired);
+
+export enum CheckoutStep {
+  Address = 1,
+  Shipping,
+  Payment,
+  Review,
+  PaymentConfirm,
+}
+
+export interface CheckoutStepDefinition {
+  index: number;
+  link: string;
+  name: string;
+  step: CheckoutStep;
+  nextActionName?: string;
+  onlyIfShippingRequired?: boolean;
+  withoutOwnView?: boolean;
+}
+
+export const CHECKOUT_STEPS: CheckoutStepDefinition[] = [
+  {
+    index: 0,
+    link: paths.checkoutAddress,
+    name: "Address",
+    nextActionName: "Continue to Shipping",
+    onlyIfShippingRequired: false,
+    step: CheckoutStep.Address,
+  },
+  {
+    index: 1,
+    link: paths.checkoutShipping,
+    name: "Shipping",
+    nextActionName: "Continue to Payment",
+    onlyIfShippingRequired: true,
+    step: CheckoutStep.Shipping,
+  },
+  {
+    index: 2,
+    link: paths.checkoutPayment,
+    name: "Payment",
+    nextActionName: "Continue to Review",
+    onlyIfShippingRequired: false,
+    step: CheckoutStep.Payment,
+  },
+  {
+    index: 3,
+    link: paths.checkoutReview,
+    name: "Review",
+    nextActionName: "Place order",
+    onlyIfShippingRequired: false,
+    step: CheckoutStep.Review,
+  },
+  {
+    index: 4,
+    link: paths.checkoutPaymentConfirm,
+    name: "Payment confirm",
+    onlyIfShippingRequired: false,
+    step: CheckoutStep.PaymentConfirm,
+    withoutOwnView: true,
+  },
+];
 
 export type SubpageCompleteHandler = () => void | Promise<void>;
