+++ src/@next/pages/ThankYouPage/ThankYouPage.tsx
@@ -2,12 +2,13 @@ import React from "react";
 import { NextPage } from "next";
 
 import { ThankYou } from "@components/organisms";
-import { NotFound } from "@temp/components";
 import { paths } from "@paths";
 import { useAuth } from "@saleor/sdk";
+
+import NotFound from "@temp/components/NotFound";
 import { IProps } from "./types";
 
-const ThankYouPage: NextPage<IProps> = ({
+export const ThankYouPage: NextPage<IProps> = ({
   query: { orderNumber, token, orderStatus },
 }) => {
   const { user } = useAuth();
@@ -26,5 +27,3 @@ const ThankYouPage: NextPage<IProps> = ({
     <NotFound />
   );
 };
-
-export { ThankYouPage };
