+++ src/core/config.ts
@@ -1,7 +1,6 @@
 /* eslint-disable global-require */
 
 import { ssrMode } from "@temp/constants";
-import { paths } from "@paths";
 import { generatePageUrl } from "./utils";
 
 export const PRODUCTS_PER_PAGE = 6;
@@ -73,63 +72,3 @@ export const META_DEFAULTS = {
   type: "website",
   url: !ssrMode ? window.location.origin : "",
 };
-export enum CheckoutStep {
-  Address = 1,
-  Shipping,
-  Payment,
-  Review,
-  PaymentConfirm,
-}
-
-export interface CheckoutStepDefinition {
-  index: number;
-  link: string;
-  name: string;
-  step: CheckoutStep;
-  nextActionName?: string;
-  onlyIfShippingRequired?: boolean;
-  withoutOwnView?: boolean;
-}
-
-export const CHECKOUT_STEPS: CheckoutStepDefinition[] = [
-  {
-    index: 0,
-    link: paths.checkoutAddress,
-    name: "Address",
-    nextActionName: "Continue to Shipping",
-    onlyIfShippingRequired: false,
-    step: CheckoutStep.Address,
-  },
-  {
-    index: 1,
-    link: paths.checkoutShipping,
-    name: "Shipping",
-    nextActionName: "Continue to Payment",
-    onlyIfShippingRequired: true,
-    step: CheckoutStep.Shipping,
-  },
-  {
-    index: 2,
-    link: paths.checkoutPayment,
-    name: "Payment",
-    nextActionName: "Continue to Review",
-    onlyIfShippingRequired: false,
-    step: CheckoutStep.Payment,
-  },
-  {
-    index: 3,
-    link: paths.checkoutReview,
-    name: "Review",
-    nextActionName: "Place order",
-    onlyIfShippingRequired: false,
-    step: CheckoutStep.Review,
-  },
-  {
-    index: 4,
-    link: paths.checkoutPaymentConfirm,
-    name: "Payment confirm",
-    onlyIfShippingRequired: false,
-    step: CheckoutStep.PaymentConfirm,
-    withoutOwnView: true,
-  },
-];
